revert merge release-43
