+++ src/components/Synthetics/TableMarketFilter/MarketFilterBase.tsx
@@ -40,10 +40,7 @@ export function MarketFilterBase({
 }: MarketFilterBaseProps) {
   const marketsInfoData = useMarketsInfoData();
   const chainId = useSelector(selectChainId);
-  const { marketTokensData: depositMarketTokensData } = useMarketTokensData(chainId, {
-    isDeposit: true,
-    withGlv: false,
-  });
+  const { marketTokensData: depositMarketTokensData } = useMarketTokensData(chainId, { isDeposit: true });
   const { marketsInfo: markets } = useSortedPoolsWithIndexToken(marketsInfoData, depositMarketTokensData);
 
   const marketsOptions = useMemo<Item<string>[]>(() => {
