+++ src/components/Synthetics/TableMarketFilter/MarketFilterLongShort.tsx
@@ -49,10 +49,7 @@ export function MarketFilterLongShort({ value, onChange, withPositions, asButton
   const marketsInfoData = useMarketsInfoData();
   const allPositions = useSelector(selectPositionsInfoDataSortedByMarket);
   const filteredPositions = useSelector(selectPositionsWithOrders);
-  const { marketTokensData: depositMarketTokensData } = useMarketTokensData(chainId, {
-    isDeposit: true,
-    withGlv: false,
-  });
+  const { marketTokensData: depositMarketTokensData } = useMarketTokensData(chainId, { isDeposit: true });
   const { marketsInfo: allMarkets } = useSortedPoolsWithIndexToken(marketsInfoData, depositMarketTokensData);
 
   const marketsOptions = useMemo<Group<MarketFilterLongShortItemData>[]>(() => {
