+++ src/context/SyntheticsStateContext/SyntheticsStateContextProvider.tsx
@@ -151,7 +151,6 @@ export function SyntheticsStateContextProvider({
     isDeposit: true,
     account,
     glvData: glvInfo.glvData,
-    withGlv: shouldFetchGlvMarkets,
   });
   const { positionsConstants } = usePositionsConstantsRequest(chainId);
   const { uiFeeFactor } = useUiFeeFactorRequest(chainId);
