+++ src/domain/stats/useFeesSummary.js
@@ -1,18 +1,10 @@
 import { getServerUrl } from "config/backend";
-import { ARBITRUM, AVALANCHE } from "config/chains";
+import { ARBITRUM, AVALANCHE, AVALANCHE_FUJI } from "config/chains";
 import { arrayURLFetcher } from "lib/fetcher";
 import useSWR from "swr";
+const ACTIVE_CHAIN_IDS = [ARBITRUM, AVALANCHE, AVALANCHE_FUJI];
 
-const ACTIVE_CHAIN_IDS = [ARBITRUM, AVALANCHE];
-
-export function useFeesSummary(): {
-  data: {
-    [chainId: string]: {
-      lastUpdatedAt: number;
-      totalFees: number;
-    };
-  };
-} {
+export function useFeesSummary() {
   const { data: feesSummary } = useSWR(
     ACTIVE_CHAIN_IDS.map((chainId) => getServerUrl(chainId, "/fees_summary")),
     {
