+++ src/domain/synthetics/markets/useMarketTokensData.ts
@@ -27,23 +27,15 @@ type MarketTokensDataResult = {
 
 export function useMarketTokensDataRequest(
   chainId: number,
-  p: {
-    isDeposit: boolean;
-    account?: string;
-    glvData?: GlvInfoData;
-    /**
-     * @default true
-     */
-    withGlv?: boolean;
-  }
+  p: { isDeposit: boolean; account?: string; glvData?: GlvInfoData; withGlv?: boolean }
 ): MarketTokensDataResult {
-  const { isDeposit, account, glvData = {}, withGlv = true } = p;
+  const { isDeposit, account, glvData = {} } = p;
   const { tokensData } = useTokensDataRequest(chainId);
   const { marketsData, marketsAddresses } = useMarkets(chainId);
 
   let isGlvTokensLoaded;
 
-  if (withGlv === false) {
+  if (p.withGlv === false) {
     isGlvTokensLoaded = true;
   } else {
     isGlvTokensLoaded = isGlvEnabled(chainId) ? Object.values(glvData).length > 0 : true;
