--- src/pages/Dashboard/getWhitelistedTokenAddresses.tsx
@@ -1,6 +1,0 @@
-import { getWhitelistedV1Tokens } from "config/tokens";
-
-export function getWhitelistedTokenAddresses(chainId) {
-  const whitelistedTokens = getWhitelistedV1Tokens(chainId);
-  return whitelistedTokens.map((token) => token.address);
-}
